Replace install-plugins.sh, append Java opts w/TLS
The emailext plugin uses a different set of configurations vs the normal
email notifications in Jenkins. The configuration for using TLS with
the SMTP server appears to be non-existant but stack suggests this
can be worked around by starting Java with this option for emailext.

plugin-installation-manager-tool appears to be the replacement for
install-plugins.sh, so my hope is to give this a whirl.