Add unit tests to eventually integrate into ci
I chose to use the unittest framework that is part of the standard
library to write tests in because it is one less dependency to track
via poetry (vs pytest), the test themselves are straight forward despite
others mentioning that unittest's framework has tests require more
boilerplate, and we already are performing setup for a python virtual
environment.

Another consideration was https://github.com/bats-core/bats-core,
however, aside the reasons above, unittest will have more support going
forward and I typically should not lean on shell scripting for things
that could grow in complexity because of the implementation language.