Add test gh-actions workflow to run tests
The 'Test' workflow only has a single generic job associated with it.
This is ok for now, if additional testing (and is different from just
testing the Makefile's targets) is required, then I intend to specify
the job names with more granularity.

Something along the lines like "test-makefile-targets". This should
take the workflow name and append it as a prefix. The workflow name
perhaps should stay as a verb of the form (e.g. Test, Publish, Build)?