Fix versioning logic and semantics
Turns out that the Jenkins versioning does not actually follow semantic
versioning. This was discovered as the Jenkins version sometimes omits
the patch versioning all together. Hence, I've changed the versioning
logic and meaning throughout the codebase.

The main regex for the versioning now allows omittance of the patch
versioning.