Update renovate schedule and packageRules
The original packageRules would have applied the same strategy for all
package types. Which wouldn't be what I want to see. The only thing
changing from the Dockerfile should be the digest. For poetry, mainly
because major updates to it dependencies could break CI/CD for autotag
I think the better strategy should be minor and patch. As for
gh-actions, the main action I am concerned with would be checkout but
even then, I don't see that breaking CI/CD.

The schedule should run prior to any in house CI/CD I plan on having.