Add explicit versioning for the Jenkins plugins
These versions should be the latest for each plugin. This was done as I
want to make sure plugins are not waiting on other versioning to
increment so the newer image can receive updates (if any) to its
associated plugins.

No matter what plugin is updated or type of version update, I only want
the container to increment it's patch versioning. Mainly as I do not see
any of these plugins entirely bricking Jenkins.

The versioning updates for the Jenkins plugins will now be handled by
Renovate.