Reseat latest tag on Jenkins plugin updates
I noticed that Renovate will update the Jenkins plugin file by doing one
plugin at a time. Meaning, the current implementation would increment
the patch number frequently.

While it maybe handy to have 10+ or more images that may only differ by
a Jenkins plugin (patch) update. I see that being a lot of waste
generated and the original intent of explicity versioning the
plugins was to ensure they would not be dependent on
external changes to have them also be updated.

Hence, if there is no version update but the Jenkins plugins file does
change, then I want to ensure that the latest tag does get reseated and
that a new image of that version gets built.

In the event a plugin breaks, one idea is to create a separate branch,
change the version of that plugin that broke, and build a custom image
off that branch. Of course, in the event that actually happens, I can
think of other ideas to get around the issue.