Refactor github workflow files
In yaml files, all members in a list (at least, this is my personal
style) should have a dash and space character as a prefix. If this list
is tied to a key, then I do expect all members of said list to be
indented by a tab.

Using DOCKERHUB_USERNAME as a secret causes my named actions to have
'*'s as part of the action name (e.g. ***/pyenv-setup@v1.0.0). Which is
well, odd, so hence my username will no longer be a secret.

WORKFLOW_ACCESS_TOKEN is now WORKFLOW_ACCESS_GH_PAT. This was done
because I want the environment variable for this secret to better
describe what  it is. Particularly it is a GitHub PAT (personal access
token) meant for being able to trigger workflows on github actions.