Use a different mirror for downloading plugins
Overtime I have noticed that many test workflows are failing only
because when making an image, the upstream plugin server sometimes
disconnects when the image tries to gather its desired plugins, causing
a workflow to fail.

Hence, I want to try and see if using a non-default mirror helps improve
the situtation. This could also be because of the large number of
concurrent connections (via CI/CD) to the default plugin server, but
for the moment I will just change the plugin server to pull from.