Migrate repo to use autotag framework
Essentially a lot of what was originally in autotag.py has been moved
to a new repository for reuseability in other projects. The client in
this case only needs to determine update types from changes made from
the prior commit. Giving those to autotag, then tags will be adjusted
and pushed.