Move the ansible-lint config into ./.config
At the time of this writing ansible-lint supports either .ansible-lint
or ./.config/ansible-lint.yml as a configuration file for ansible-lint.
The latter I feel is a better description of the type of configuration
file (e.g. json vs yaml). Hence the transition and what I believe it is
what I wanted originally.